new 提交订单

+++ mall-entity/src/main/java/com/gt/mall/entity/order/MallOrderReturn.java
@@ -152,7 +152,7 @@ public class MallOrderReturn extends Model< MallOrderReturn > {
      * 退还积分
      */
     @TableField( "return_jifen" )
-    private Integer    returnJifen;
+    private Double     returnJifen;
 
     @Override
     protected Serializable pkVal() {
